+++ utils/js/LangUtils.ts
@@ -4,7 +4,7 @@
  *
  * Copyright © 2024 Extremely Heavy Industries Inc.
  */
-import {Thunkable, XH} from '@xh/hoist/core';
+import {Thunkable} from '@xh/hoist/core';
 import {Exception} from '@xh/hoist/core/exception/Exception';
 import {
     flatMap,
@@ -19,7 +19,6 @@ import {
     uniq,
     uniqBy
 } from 'lodash';
-import {v4} from 'uuid';
 import _inflection from 'lodash-inflection';
 
 mixin(_inflection);
@@ -301,11 +300,4 @@ export function intersperse<T>(arr: T[], separator: T): T[] {
  */
 export function executeIfFunction<T>(v: Thunkable<T>): T {
     return isFunction(v) ? v() : v;
-}
-
-/**
- * Generate a UUID, prefixed with the application code.
- */
-export function genUUID(): string {
-    return `${XH.appCode}-${v4()}`;
 }