+++ CHANGELOG.md
@@ -8,13 +8,14 @@
 * Added support for Correlation ID's across fetch requests and error / activity tracking:
      * Requires `hoist-core >= v19.0.0`.
      * Correlation ID's are exposed in `Admin` > `Activity` > `Tracking` and `Client Errors` tabs.
-     * `FetchService` now has `setCorrelationIdToken` and `setGenerateCorrelationIdsByDefault`
-        methods for specifying a Correlation ID header key and globally enabling Correlation ID
-        generation unless explicitly overridden via the following configs:
+     * `FetchService` now has `setCorrelationIdHeaderKey` and `enableCorrelationIds` methods for
+        specifying a Correlation ID header key and globally enabling Correlation ID generation
+        unless explicitly overridden via the following configs:
      * `TrackOptions.correlationId` - specify a Correlation ID for a tracked activity
      * `LoadSpec.correlationId` - specify a Correlation ID or set to `true` to auto-generate
      * `FetchOptions.correlationId` - specify a Correlation ID. Set to `true` to source from
         provided `LoadSpec` or auto-generate. Set to `false` to explicitly omit.
+     * New `FetchService.generateCorrelationId` method for manual Correlation ID generation.
 
 ### 🐞 Bug Fixes